refactor(manager): 466796 - Sonarlint - refactor Service
