+++ src/api/forms/service/definition.js
@@ -1,11 +1,6 @@
-import {
-  ControllerType,
-  formDefinitionSchema,
-  slugify
-} from '@defra/forms-model'
+import { formDefinitionSchema, slugify } from '@defra/forms-model'
 import Boom from '@hapi/boom'
 import { MongoServerError } from 'mongodb'
-import { v4 as uuidV4 } from 'uuid'
 
 import {
   makeFormLiveErrorMessages,
@@ -15,28 +10,15 @@ import { InvalidFormDefinitionError } from '~/src/api/forms/errors.js'
 import * as formDefinition from '~/src/api/forms/repositories/form-definition-repository.js'
 import * as formMetadata from '~/src/api/forms/repositories/form-metadata-repository.js'
 import {
-  findComponent,
-  findPage,
-  summaryHelper,
-  uniquePathGate
-} from '~/src/api/forms/repositories/helpers.js'
+  DRAFT,
+  defaultAuthor,
+  defaultDate,
+  logger,
+  partialAuditFields
+} from '~/src/api/forms/service/shared.js'
 import * as formTemplates from '~/src/api/forms/templates.js'
-import { createLogger } from '~/src/helpers/logging/logger.js'
 import { client } from '~/src/mongo.js'
 
-const logger = createLogger()
-
-const defaultAuthor = {
-  displayName: 'Unknown',
-  id: '-1'
-}
-
-const defaultDate = new Date('2024-06-25T23:00:00Z') // date we went live
-/**
- * @typedef {'draft' | 'live'} State
- */
-
-const DRAFT = /** @type {State} */ ('draft')
 /**
  * Maps a form metadata document from MongoDB to form metadata
  * @param {WithId<Partial<FormMetadataDocument>>} document - form metadata document (with ID)
@@ -220,22 +202,6 @@ export async function getFormBySlug(slug) {
  */
 export function getFormDefinition(formId, state = DRAFT, session = undefined) {
   return formDefinition.get(formId, state, session)
-}
-
-/**
- * Partially update the [state] fields
- * @param {Date} date
- * @param {FormMetadataAuthor} author
- * @param {string} state
- * @returns {PartialFormMetadataDocument}
- */
-const partialAuditFields = (date, author, state = DRAFT) => {
-  return /** @type {PartialFormMetadataDocument} */ {
-    [`${state}.updatedAt`]: date,
-    [`${state}.updatedBy`]: author,
-    updatedAt: date,
-    updatedBy: author
-  }
 }
 
 /**
@@ -530,428 +496,6 @@ export async function removeForm(formId) {
   logger.info(`Removed form with ID ${formId}`)
 }
 
-const SUMMARY_PAGE_ID = '449a45f6-4541-4a46-91bd-8b8931b07b50'
-/**
- * Adds id to summary page if it's missing
- * @param {PageSummary} summaryPage
- */
-export const addIdToSummary = (summaryPage) => ({
-  id: SUMMARY_PAGE_ID,
-  ...summaryPage
-})
-
-/**
- * Repositions the summary page if it's not the last index of pages
- * @param {string} formId
- * @param {FormDefinition} definition
- * @param {FormMetadataAuthor} author
- */
-export async function repositionSummaryPipeline(formId, definition, author) {
-  const summaryResult = summaryHelper(definition)
-  const { shouldRepositionSummary } = summaryResult
-
-  logger.info(`Checking position of summary on ${formId}`)
-
-  if (!shouldRepositionSummary) {
-    logger.info(`Position of summary on ${formId} correct`)
-    return summaryResult
-  }
-
-  logger.info(`Updating position of summary on Form ID ${formId}`)
-
-  const session = client.startSession()
-
-  const { summary } = summaryResult
-  const summaryDefined = /** @type {PageSummary} */ (summary)
-  const summaryWithId = addIdToSummary(summaryDefined)
-
-  try {
-    await session.withTransaction(async () => {
-      await formDefinition.removeMatchingPages(
-        formId,
-        { controller: ControllerType.Summary },
-        session
-      )
-
-      await formDefinition.addPageAtPosition(
-        formId,
-        /** @type {PageSummary} */ (summaryWithId),
-        session,
-        {}
-      )
-
-      // Update the form with the new draft state
-      await formMetadata.update(
-        formId,
-        { $set: partialAuditFields(new Date(), author) },
-        session
-      )
-    })
-  } catch (err) {
-    logger.error(
-      err,
-      `Failed to update position of summary on Form ID ${formId}`
-    )
-    throw err
-  } finally {
-    await session.endSession()
-  }
-
-  logger.info(`Updated position of summary on Form ID ${formId}`)
-
-  return { ...summaryResult, summary: summaryWithId }
-}
-
-/**
- * Adds an id to a page
- * @param {Page} pageWithoutId
- * @returns {Page}
- */
-const createPageWithId = (pageWithoutId) => ({
-  ...pageWithoutId,
-  id: uuidV4().toString()
-})
-
-/**
- * Adds a new page to a draft definition and calls repositionSummaryPipeline is summary exists
- * @param {string} formId
- * @param {Page} newPage
- * @param {FormMetadataAuthor} author
- */
-export async function createPageOnDraftDefinition(formId, newPage, author) {
-  logger.info(`Creating new page for form with ID ${formId}`)
-
-  const session = client.startSession()
-
-  /** @type {FormDefinition} */
-  const formDraftDefinition = await getFormDefinition(formId, DRAFT)
-
-  uniquePathGate(
-    formDraftDefinition,
-    newPage.path,
-    `Duplicate page path on Form ID ${formId}`
-  )
-
-  const { summaryExists } = await repositionSummaryPipeline(
-    formId,
-    formDraftDefinition,
-    author
-  )
-  /**
-   * @type {{ position?: number; state?: 'live' | 'draft' }}
-   */
-  const options = {}
-
-  if (summaryExists) {
-    options.position = -1
-  }
-
-  const page = createPageWithId(newPage)
-
-  try {
-    await session.withTransaction(async () => {
-      await formDefinition.addPageAtPosition(formId, page, session, options)
-
-      // Update the form with the new draft state
-      await formMetadata.update(
-        formId,
-        { $set: partialAuditFields(new Date(), author) },
-        session
-      )
-    })
-  } catch (err) {
-    logger.error(err, `Failed to add page on ${formId}`)
-    throw err
-  } finally {
-    await session.endSession()
-  }
-
-  logger.info(`Created new page for form with ID ${formId}`)
-
-  return page
-}
-
-/**
- * Adds id to a component
- * @param {ComponentDef} component
- * @returns {ComponentDef}
- */
-const addIdToComponent = (component) =>
-  /** @type {ComponentDef} */ ({
-    ...component,
-    id: uuidV4()
-  })
-
-/**
- * Gets a page from a form definition and fails if the page is not found
- * @param {string} formId
- * @param {string} pageId
- * @param {ClientSession} [session]
- */
-export async function getFormDefinitionPage(formId, pageId, session) {
-  logger.info(`Getting Page ID ${pageId} on Form ID ${formId}`)
-
-  const definition = /** @type {FormDefinition} */ await getFormDefinition(
-    formId,
-    DRAFT,
-    session
-  )
-
-  const page = findPage(definition, pageId)
-
-  if (page === undefined) {
-    throw Boom.notFound(`Page ID ${pageId} not found on Form ID ${formId}`)
-  }
-
-  logger.info(`Got Page ID ${pageId} on Form ID ${formId}`)
-
-  return page
-}
-
-/**
- * Gets a component from a formDefintion page if it exists, throws a Boom.notFound if not
- * @param {string} formId
- * @param {string} pageId
- * @param {string} componentId
- * @param {ClientSession} [session]
- */
-export async function getFormDefinitionPageComponent(
-  formId,
-  pageId,
-  componentId,
-  session
-) {
-  logger.info(
-    `Getting Component ID ${componentId} on Page ID ${pageId} & Form ID ${formId}`
-  )
-
-  const definition = /** @type {FormDefinition} */ await getFormDefinition(
-    formId,
-    DRAFT,
-    session
-  )
-
-  const component = findComponent(definition, pageId, componentId)
-
-  if (component === undefined) {
-    throw Boom.notFound(
-      `Component ID ${componentId} not found on Page ID ${pageId} & Form ID ${formId}`
-    )
-  }
-  logger.info(
-    `Got Component ID ${componentId} on Page ID ${pageId} & Form ID ${formId}`
-  )
-
-  return component
-}
-/**
- * Adds a component to the end of page components
- * @param {string} formId
- * @param {string} pageId
- * @param {ComponentDef[]} components
- * @param {FormMetadataAuthor} author
- * @param {boolean} prepend
- */
-export async function createComponentOnDraftDefinition(
-  formId,
-  pageId,
-  components,
-  author,
-  prepend = false
-) {
-  await getFormDefinitionPage(formId, pageId)
-
-  logger.info(`Adding new component on Page ID ${pageId} on Form ID ${formId}`)
-
-  const session = client.startSession()
-
-  const createdComponents =
-    /** @type {ComponentDef[]} */ components.map(addIdToComponent)
-
-  const positionOptions = /** @satisfies {{ position?: number }} */ {}
-
-  if (prepend) {
-    positionOptions.position = 0
-  }
-
-  try {
-    await session.withTransaction(async () => {
-      await formDefinition.addComponents(
-        formId,
-        pageId,
-        createdComponents,
-        session,
-        { state: DRAFT, ...positionOptions }
-      )
-
-      // Update the form with the new draft state
-      await formMetadata.update(
-        formId,
-        { $set: partialAuditFields(new Date(), author) },
-        session
-      )
-    })
-  } catch (err) {
-    logger.error(
-      err,
-      `Failed to add component on Page ID ${pageId} Form ID ${formId}`
-    )
-    throw err
-  } finally {
-    await session.endSession()
-  }
-
-  logger.info(`Added new component on Page ID ${pageId} on Form ID ${formId}`)
-
-  return createdComponents
-}
-
-/**
- * Updates a component and throws a Boom.notFound if page or component is not found
- * @param {string} formId
- * @param {string} pageId
- * @param {string} componentId
- * @param {ComponentDef} component
- * @param {FormMetadataAuthor} author
- */
-export async function updateComponentOnDraftDefinition(
-  formId,
-  pageId,
-  componentId,
-  component,
-  author
-) {
-  logger.info(
-    `Updating Component ID ${componentId} on Page ID ${pageId} & Form ID ${formId}`
-  )
-
-  let componentReturn = await getFormDefinitionPageComponent(
-    formId,
-    pageId,
-    componentId
-  )
-
-  const session = client.startSession()
-
-  try {
-    await session.withTransaction(
-      async () => {
-        await formDefinition.updateComponent(
-          formId,
-          pageId,
-          componentId,
-          component,
-          session,
-          DRAFT
-        )
-
-        componentReturn = await getFormDefinitionPageComponent(
-          formId,
-          pageId,
-          componentId,
-          session
-        )
-
-        // Check that component has been updated
-        if (JSON.stringify(componentReturn) !== JSON.stringify(component)) {
-          throw Boom.internal(
-            `Component ${componentId} not updated on Page ID ${pageId} and Form ID ${formId}`
-          )
-        }
-
-        // Update the form with the new draft state
-        await formMetadata.update(
-          formId,
-          { $set: partialAuditFields(new Date(), author) },
-          session
-        )
-      },
-      { readPreference: 'primary' }
-    )
-  } catch (err) {
-    logger.error(
-      err,
-      `Failed to update Component ID ${componentId} on Page ID ${pageId} & Form ID ${formId}`
-    )
-    throw err
-  } finally {
-    await session.endSession()
-  }
-
-  logger.info(
-    `Updated Component ID ${componentId} on Page ID ${pageId} & Form ID ${formId}`
-  )
-
-  return componentReturn
-}
-/**
- * Updates specific fields on a page, allowing concurrent changes should components be updated
- * @param {string} formId
- * @param {string} pageId
- * @param {PatchPageFields} pageFieldsToUpdate
- * @param {FormMetadataAuthor} author
- */
-export async function patchFieldsOnDraftDefinitionPage(
-  formId,
-  pageId,
-  pageFieldsToUpdate,
-  author
-) {
-  let page = await getFormDefinitionPage(formId, pageId)
-
-  const session = client.startSession()
-  const fields = /** @type {(keyof PatchPageFields)[]} */ (
-    Object.keys(pageFieldsToUpdate)
-  )
-
-  try {
-    await session.withTransaction(
-      async () => {
-        await formDefinition.updatePageFields(
-          formId,
-          pageId,
-          pageFieldsToUpdate,
-          session,
-          DRAFT
-        )
-
-        page = await getFormDefinitionPage(formId, pageId, session)
-
-        // Check whether field changes have persisted and abort transaction if not
-        const failedFields = /** @type {(keyof PatchPageFields)[]} */ ([])
-
-        fields.forEach((field) => {
-          if (page[field] !== pageFieldsToUpdate[field]) {
-            failedFields.push(field)
-          }
-        })
-        if (failedFields.length) {
-          throw Boom.internal(
-            `Failed to patch fields ${failedFields.toString()} on Page ID ${pageId} Form ID ${formId}`
-          )
-        }
-
-        // Update the form with the new draft state
-        await formMetadata.update(
-          formId,
-          { $set: partialAuditFields(new Date(), author) },
-          session
-        )
-      },
-      { readPreference: 'primary' }
-    )
-  } catch (err) {
-    logger.error(
-      err,
-      `Failed to patch fields ${fields.toString()} on Page ID ${pageId} Form ID ${formId}`
-    )
-    throw err
-  } finally {
-    await session.endSession()
-  }
-
-  return page
-}
 /**
  * @import { FormDefinition, FormMetadataAuthor, FormMetadataDocument, FormMetadataInput, FormMetadata, FilterOptions, QueryOptions, Page, PageSummary, FormStatus, ComponentDef, PatchPageFields } from '@defra/forms-model'
  * @import { WithId, UpdateFilter, ClientSession } from 'mongodb'
